Nextjs page to app
