+++ package.json
@@ -10,11 +10,15 @@
     "lint": "next lint --fix",
     "lint:check": "next lint"
   },
+  "engines": {
+    "node": "22.x"
+  },
   "dependencies": {
     "@emotion/react": "^11.8.1",
     "@emotion/styled": "^11.6.0",
     "@mui/icons-material": "^7.0.2",
     "@mui/material": "^7.0.2",
+    "@mui/material-nextjs": "^7.0.2",
     "@mui/x-date-pickers": "^7.28.3",
     "@next/third-parties": "15.3.0",
     "@syncfusion/ej2-base": "^29.1.36",
@@ -24,8 +28,8 @@
     "dotenv": "^16.0.0",
     "next": "15.3.0",
     "postcss": "^8.5.3",
-    "react": "18.3.1",
-    "react-dom": "18.3.1",
+    "react": "19.1.0",
+    "react-dom": "19.1.0",
     "sharp": "^0.34.1",
     "tailwindcss": "^4.1.3"
   },
@@ -34,12 +38,12 @@
     "@types/react": "19.1.0",
     "@typescript-eslint/eslint-plugin": "^8.29.1",
     "@typescript-eslint/parser": "^8.29.1",
-    "eslint": "^8.57.1",
+    "eslint": "^9.24.0",
     "eslint-config-next": "15.3.0",
-    "eslint-config-prettier": "^9.1.0",
+    "eslint-config-prettier": "^10.1.1",
     "eslint-plugin-jsx-a11y": "^6.5.1",
     "eslint-plugin-react": "^7.37.5",
-    "eslint-plugin-react-hooks": "^4.3.0",
+    "eslint-plugin-react-hooks": "^5.2.0",
     "eslint-plugin-simple-import-sort": "^12.1.1",
     "prettier": "^3.5.3",
     "typescript": "^5.8.3"
