+++ src/app/loading.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+import Home from './Home';
+
+/**
+ * Returns the home page with Nebula Branding and search options
+ */
+export default function Loading() {
+  return <Home roomsLoading />;
+}
