+++ src/app/results/opengraph-image.alt.txt
@@ -1,0 +1,1 @@
+Nebula Labs Icon.
