+++ src/app/robots.txt
@@ -1,0 +1,4 @@
+User-Agent: *
+Allow: /
+
+Sitemap: https://rooms.utdnebula.com/sitemap.xml
