+++ src/lib/localization.tsx
@@ -1,0 +1,17 @@
+'use client';
+
+import { LocalizationProvider } from '@mui/x-date-pickers';
+import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs';
+import React from 'react';
+
+export default function ClientLocalizationProvider({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <LocalizationProvider dateAdapter={AdapterDayjs}>
+      {children}
+    </LocalizationProvider>
+  );
+}
