+++ tsconfig.json
@@ -18,8 +18,20 @@
       "@/*": ["*"]
     },
     "baseUrl": "src",
-    "verbatimModuleSyntax": true
+    "verbatimModuleSyntax": true,
+    "plugins": [
+      {
+        "name": "next"
+      }
+    ]
   },
-  "include": ["src", "next-env.d.ts", "**/*.ts", "**/*.tsx", "/**/**/*/.tsx"],
+  "include": [
+    "**/*.ts",
+    "**/*.tsx",
+    "/**/**/*/.tsx",
+    "next-env.d.ts",
+    "src",
+    ".next/types/**/*.ts"
+  ],
   "exclude": ["node_modules"]
 }
